Ensure the background is black and the foreground is white for the current turn circle in the Interactive#game_header
This is important because the circle symbols used (●, ○) assume that the
background is black and the foreground is white. So white's turn shows
the full circle filled in as white, and the "ring" circle is just a
white outline.

If someone's terminal was white on black this will have the opposite
effect than a black on white (eg: a black filled in circle with the
label "white" would be confusing). To avoid this confusion, the
background and foreground are set explicitly.